Update patch set 1
Patch Set 1:

I was wondering where franco got his patches from.

As far as the other bcmdhd patches - Those are as important as this one.  Their impact is not minor, standby drain on my device went from 20-30% battery loss overnight down to 3-4%.

We should be careful not to pull in too many patches, but the bcmdhd set has clear and significant power management benefits and this one fixes a CVE.

Patch-set: 1