Update patch set 1
Patch Set 1:

I saw it in either m7 or oppo kernels so I thought it would be ok.

Also, please consider the extreme throttling on hammerhead. Throttling even kicks in during boot and reducing voltage greatly helps against temperature. We paid for a 2.26GHz device, not a 1.5GHz one. I'd even consider to push a commit for this that raises throttling limits at least a bit.

This is not fsync and not overclocking/-volting. If a user chooses too low voltages, the device just dies - end of story. Everybody is aware of that. Voltage limits are within regular boundaries as well.

And do we really want users to use custom kernels? Wasn't it one of Omni's main goals to be user oriented?

Undervolting is _not_ dangerous.

Patch-set: 1